requirements.txt
streamlit 
langchain
google-generativeai
python-dotenv
langchain-google-genai